pyproject.toml
[tool.poetry]
name = "qoqolodownloader"
version = "0.1.0"
description = ""
authors = ["jiade <jiade@jiade.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
requests = "^2.32.3"
pathvalidate = "^3.2.1"
piexif = "^1.1.3"
jproperties = "^2.1.2"
selenium = "^4.27.1"
python-dateutil = "^2.9.0.post0"
pyinstaller = "^6.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
build = "qoqolodownloader.pyinstaller:install"